added change_name() function
Converts a string of up to 16 characters to a list of ascii-values and
copies them to the last 16 elements of msg_bytes. Unused characters get
filled up with spaces.

Also using map() to turn the last 16 bytes into the program_name string instead
of iterating and appending byte by byte.