added support for simple config file (.pypodrc)
the file can be either stored in the home-directory (os.environ['HOME']) or in
the current directory. It is a simple json-file that holds the values for
MIDI in- and outputs as well as the channel. This makes using the gui a little
bit more convenient since you don't have to change those values only once and as
long as you are not changing anything MIDI-wise you don't have to worry about it
anymore. Saving the config file (maybe with additional options) will be
available in the GUI as well.