"fixed" delay times
According to https://medias.audiofanzine.com/files/lin020-477344.pdf the delay time should be saved in bytes 27-30, but even with the delay time at maximum I only get a 1 in bit 28. When I dump a program and look at the lower 7 bits of bytes 29 and 30 I get almost exactly the values that I set before saving the program (by sending MIDI CC 30 and 62). But I am still not sure if this is the correct way since the above mentioned document says something about 17bit. At least now we get _almost_ the right values...