pypod_gui.py: yet another delay time update
I am still not convinced that this is not really the right way to get the delay values from the sysex data but at least I always get pretty much the same values back when I change them in the gui and then immediately download the edit buffer. Only the first delay time is sometimes off by one or two which is probably caused by the fact that again we only get 6 bits for a 7bit-number so we have to multiply by 2. This will probably not be the last update to this part but for now I will keep it as it is before I go completely mad ;)