changed most Switch-widgets to GtkCheckButtons
A GtkSwitch would not always return the correct state which gave incorrect results
when downloading/loading programs (Switch turned on but effect turned off etc.).
CheckButtons are not so nice looking but they seem to work. I will keep Switches for
Modulation, Wah and Volume Swell since those are settings that are made to be used
with a foot switch and not to be saved in the program anyway. If I find any problems
with those they will probably also be changed.