TODO: disable signal handlers during updateGUI()
as described in the readme, this leads to re-sending the exact same information to the pod that we
just downloaded, makes no sense and always marks the program as changed, although this is not the
case (at least if we download the currently selected program). So at the beginning of updateGUI()
all signals should be ignored and at the end of the function they should be reactivated again.
This is probably possible with something like signal_handler_block / signal_handler_unblock
(https://lazka.github.io/pgi-docs/#GObject-2.0/functions.html%23GObject.signal_handler_block)
This is mainly a problem when sending program changes to the pod, not so much when downloading a
program since we probably want to change it anyway and if we have another program selected this
is the right behaviour (e.g. have 4D selected, download 5A, 4D will still be selected but marked
as changed which it probably is and in this case we even want updateGUI to also update the pod's
edit buffer).