pypod_gui.py: removed sleep and updateGUI from program_change()
after thinking about it, I think it makes not much sense to call updateGUI() after
sending a program change. Although it would be nice to always have GUI and pod
synchronized, there is no real need for it. If we want the GUI represent
the current settings on the pod, dump program or dump edit buffer is the right
option.