GUI: various fixes, removed some time.sleep() calls
there are only two time.sleep() calls left: when saving a file and
when uploading a program. At the moment I do not see a good way to get rid
of those. But otherwise the GUI responds better because as soon as it
receives data it calls updateGUI(). There is still a little bug when switching
programs that seems to be caused by the "Modulation"-Switch. Somehow this confuses
the pod and when the Switch is off and we load a new program that uses a
modulation effect it gets disabled and the program is marked as changed.